requirements.txt
pyspark
pandas
numpy
scipy
scikit-learn
polars
orjson
pyarrow
awswrangler
transformers
accelerate
duckdb
neo4j
s3fs
umap-learn
smart-open
onnxruntime
spacy
seqeval
gensim
numba
sqlalchemy
pytest
pyyaml
datasets
jsonschema>=3.0
rsa<4.8,>=3.1.2
cachetools<6,>=4.0
gitpython!=3.1.19,<4,>=3.0.7
pillow<12,>=7.1.0
tenacity<10,>=8.1.0
toml<2,>=0.10.1
tornado<7,>=6.0.3
watchdog<7,>=2.1.5
botocore<1.36.24
